Support for new "--coverage-exclude" option.
The new option accepts a glob-like string and will exclude (from code
coverage reports) those files whose realpath is - at least
partially - matched by the specified glob.

This actually solves #225 - and adds a bit more (you can exclude globs and
not just specific directories, like it was proposed in the issue).

Example:
> --coverage-exclude some*/glob/**/pattern

In reality only a subset of glob-like pattern is supported: '*', '**' and '?'.
But these could cover most of usecases.

Internally, the glob-like pattern is converted to a regex - and when
we process list of files fetched by the code-coverage engine we filter
some out - based on the regex.

- The '--coverage-exclude ...' option can be specified multiple times.
- HTML template has been updated to list these excluded files at the
  bottom of the page.
- When using pcov for gathering code coverage, we even pass the resulting
  regex into 'pcov.exclude' INI setting, because pcov supports excluding
  filepaths from coverage this way.